+++ coolq/plugins/pretty_boy/pick_boy_utils.py
@@ -1,11 +1,13 @@
 import random
+from datetime import datetime
 
 import nonebot
 from nonebot.adapters.cqhttp import Bot
-from coolq.plugins.pretty_boy.config import CustomConfig
+
 from nonebot.adapters.cqhttp.message import Message
-from coolq.plugins.pretty_boy.redis_utils import Redis
+from .redis_utils import Redis
 from nonebot.log import logger
+from .config import CustomConfig
 
 
 async def base(group_id, bot_id):
@@ -28,7 +30,7 @@ def check_boy(boys):
     return sex, user_id
 
 
-def get_today_pretty_boy():
+async def get_today_pretty_boy():
     """
     获取今天的pretty_boy
     :return:
@@ -41,7 +43,23 @@ def get_today_pretty_boy():
     bot: Bot = nonebot.get_bots()[CustomConfig.bot_id]
     boys = await bot.call_api('get_group_member_list', **{'group_id': CustomConfig.group_id})
 
+    second = get_expire_time()
     sex, user_id = check_boy(boys)
     Redis.hset(key, "user_id", user_id)
     Redis.hset(key, "sex", sex)
+    Redis.expire(key, expire=second)
     return sex, user_id
+
+
+def get_expire_time():
+    """
+    获取当前时间距离今天结束
+    :return:
+    """
+    now = datetime.now()
+    day_string = now.strftime("%Y-%m-%d 23:59:59")
+    end_time = datetime.strptime(day_string, "%Y-%m-%d %H:%M:%S")
+    second = int((end_time - now).total_seconds())
+    if second <= 0:
+        second = 1
+    return second
